refactor: split encoding/encodingread.go — delta and byte stream split reads into separate files
